Avoid aliasing id & rev if prop names are already there
Because of obscure historical reasons, we have an `id` property in our CouchDB
documents. So, when one GET-s such a document with `curl`, they'll see something
like:

{ _id: '811b77f35fca01bae6eb9b8ad67becc4',
  _rev: '1-c70b6027c1d75917cef89710a6ef02f0',
  id: '10001',
  name: 'Sorell Hotel Aarauerhof',
  ratingStars: 3,
  ...}

However, if we `get()` such a document with cradle, our `id` value disappears,
becoming a non-enumerable property with a value of `811b77f...`. That change
helps avoiding such a situation.