Add a new cache mode 'follow'
Rather than specifying 'true' or 'false' for the 'cache' option,
specifying 'follow' will uses a changes feed to track external
modification to cached documents.  Only documents that have been
cached during normal write-through operation will be updated.

This is useful if you have a system where multiple clients may write
directly to the database without going through a single cache.

Note: This could be made more optimal for cases where not all changes
in the feed are cached as it pulls the full document content in the
change feed.  Maybe a mode 'follow-lite' could enable a lightweight
change feed followed by additional fetch if it is found that a changed
document is cached.  If, for a particular app however, changes are
likely to occur externally to documents in the cache, the existing
'include_docs' mode is probably OK.